requirements.txt
alembic==1.6.5
aniso8601==9.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
pkg-resources==0.0.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
